# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"TuDistri API","description":"API for Portoviejo Route Optimization","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/routes/recalculate":{"post":{"summary":"Recalculate Route","operationId":"recalculate_route_api_routes_recalculate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecalculateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecalculateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/products":{"get":{"summary":"Get Products","operationId":"get_products_api_products_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductListResponse"}}}}}}},"/api/sellers":{"get":{"summary":"Get Sellers","operationId":"get_sellers_api_sellers_get","parameters":[{"name":"product_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SellerListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/pois":{"get":{"summary":"Get Pois","operationId":"get_pois_api_pois_get","parameters":[{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/POIListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/routes/simulate":{"post":{"summary":"Simulate Routes","operationId":"simulate_routes_api_routes_simulate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimulationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimulationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/validation/stats":{"get":{"summary":"Get Validation Stats","operationId":"get_validation_stats_api_validation_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationStatsResponse"}}}}}}},"/api/demand/forecast":{"get":{"summary":"Get Demand Forecast","operationId":"get_demand_forecast_api_demand_forecast_get","parameters":[{"name":"product","in":"query","required":true,"schema":{"type":"string","title":"Product"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":7,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DemandForecastResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Coordinates":{"properties":{"lat":{"type":"number","title":"Lat"},"lng":{"type":"number","title":"Lng"}},"type":"object","required":["lat","lng"],"title":"Coordinates"},"DemandForecastResponse":{"properties":{"product":{"type":"string","title":"Product"},"forecast":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Forecast"}},"type":"object","required":["product","forecast"],"title":"DemandForecastResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"POI":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"category":{"type":"string","title":"Category"},"coordinates":{"$ref":"#/components/schemas/Coordinates"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"}},"type":"object","required":["id","name","category","coordinates"],"title":"POI"},"POIListResponse":{"properties":{"pois":{"items":{"$ref":"#/components/schemas/POI"},"type":"array","title":"Pois"}},"type":"object","required":["pois"],"title":"POIListResponse"},"Product":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"icon":{"type":"string","title":"Icon"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"price_per_unit":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Per Unit"},"unit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unit"}},"type":"object","required":["id","name","icon"],"title":"Product"},"ProductListResponse":{"properties":{"products":{"items":{"$ref":"#/components/schemas/Product"},"type":"array","title":"Products"}},"type":"object","required":["products"],"title":"ProductListResponse"},"RecalculateRequest":{"properties":{"current_lat":{"type":"number","title":"Current Lat"},"current_lng":{"type":"number","title":"Current Lng"},"dest_lat":{"type":"number","title":"Dest Lat"},"dest_lng":{"type":"number","title":"Dest Lng"},"event_type":{"type":"string","title":"Event Type"},"simulation_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Simulation Id"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id"},"progress":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Progress","default":0.0}},"type":"object","required":["current_lat","current_lng","dest_lat","dest_lng","event_type"],"title":"RecalculateRequest"},"RecalculateResponse":{"properties":{"route_geometry":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Route Geometry"},"distance_km":{"type":"number","title":"Distance Km"},"duration_min":{"type":"number","title":"Duration Min"},"event_applied":{"type":"string","title":"Event Applied"}},"type":"object","required":["route_geometry","distance_km","duration_min","event_applied"],"title":"RecalculateResponse"},"RouteResult":{"properties":{"seller_id":{"type":"string","title":"Seller Id"},"seller_name":{"type":"string","title":"Seller Name"},"seller_rating":{"type":"number","title":"Seller Rating","default":0},"seller_trips":{"type":"integer","title":"Seller Trips","default":0},"route_geometry":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Route Geometry"},"duration_seconds":{"type":"number","title":"Duration Seconds"},"distance_meters":{"type":"number","title":"Distance Meters"},"distance_km":{"type":"number","title":"Distance Km"},"duration_min":{"type":"number","title":"Duration Min"},"transport_cost":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Transport Cost"},"estimated_revenue":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Estimated Revenue"},"net_profit":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Net Profit"},"load_percentage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Load Percentage"},"product_image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Image"},"product_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Name"},"price_per_unit":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price Per Unit"},"freshness_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Freshness Score"},"punctuality_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Punctuality Score"},"satisfaction_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Satisfaction Score"},"simulation_state":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Simulation State"}},"type":"object","required":["seller_id","seller_name","route_geometry","duration_seconds","distance_meters","distance_km","duration_min"],"title":"RouteResult"},"Seller":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"products":{"items":{"type":"string"},"type":"array","title":"Products"},"coordinates":{"$ref":"#/components/schemas/Coordinates"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"},"trips_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Trips Count"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type","default":"distributor"}},"type":"object","required":["id","name","products","coordinates"],"title":"Seller"},"SellerListResponse":{"properties":{"sellers":{"items":{"$ref":"#/components/schemas/Seller"},"type":"array","title":"Sellers"}},"type":"object","required":["sellers"],"title":"SellerListResponse"},"SimulationMetrics":{"properties":{"revenue":{"type":"number","title":"Revenue"},"profit":{"type":"number","title":"Profit"},"distance_total":{"type":"number","title":"Distance Total"},"duration_total":{"type":"number","title":"Duration Total"},"platform_profit":{"type":"number","title":"Platform Profit"},"prediction_accuracy":{"type":"number","title":"Prediction Accuracy"},"avg_time_reduction":{"type":"number","title":"Avg Time Reduction"},"revenue_growth":{"type":"number","title":"Revenue Growth"}},"type":"object","required":["revenue","profit","distance_total","duration_total","platform_profit","prediction_accuracy","avg_time_reduction","revenue_growth"],"title":"SimulationMetrics"},"SimulationRequest":{"properties":{"user_lat":{"type":"number","title":"User Lat"},"user_lng":{"type":"number","title":"User Lng"},"product_id":{"type":"string","title":"Product Id"},"weight":{"type":"number","title":"Weight","default":1.0},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id"}},"type":"object","required":["user_lat","user_lng","product_id"],"title":"SimulationRequest"},"SimulationResponse":{"properties":{"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id"},"recommended_route":{"anyOf":[{"$ref":"#/components/schemas/RouteResult"},{"type":"null"}]},"all_routes":{"items":{"$ref":"#/components/schemas/RouteResult"},"type":"array","title":"All Routes"},"metrics":{"$ref":"#/components/schemas/SimulationMetrics"}},"type":"object","required":["all_routes","metrics"],"title":"SimulationResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"ValidationStatsResponse":{"properties":{"routing":{"additionalProperties":true,"type":"object","title":"Routing"},"simulation":{"additionalProperties":true,"type":"object","title":"Simulation"},"timestamp":{"type":"number","title":"Timestamp"}},"type":"object","required":["routing","simulation","timestamp"],"title":"ValidationStatsResponse"}}}}